pre-prueba/pyproject.toml
[project]
name = "retailco-analysis"
version = "0.1.0"
description = "Prueba de desempeño - Analítica de datos Jr para RetailCo"
authors = ["Analista Jr"]
dependencies = [
    "numpy>=2.0.2",
    "pandas>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
]
requires-python = ">=3.9"
